+++ CHANGELOG.md
@@ -12,6 +12,10 @@ versioning](http://semver.org).
 - Quiet option reduces verbosity
 - "Watch" plugin controls whether services and dependent modules are
   automatically reloaded when they change
+- ``tangelo.ensurePlugin()`` function that avoids JavaScript parsing problems
+- Service functions ``tangelo.redirect()`` and ``tangelo.internal_redirect()``
+  to allow services to redirect to other resources
+- Service function ``tangelo.file()`` to serve arbitrary files
 
 ### Changed
 - Documentation introduction is more focused; tutorials are more
@@ -26,6 +30,7 @@ versioning](http://semver.org).
 - Bundled Mongo plugin updated to use PyMongo 3.2
 
 ### Deprecated
+- ``tangelo.getPlugin()`` - use ``tangelo.ensurePlugin()`` instead
 
 ### Removed
 - "System Architecture" section in README
@@ -39,6 +44,8 @@ versioning](http://semver.org).
   error messages
 
 ### Security
+- ``Server`` response header now reads "Tangelo" instead of "CherryPy" with a
+  version string
 - Runtime exceptions no longer send tracebacks to the client; instead, an error
   report code is sent, which can be matched up to a traceback appearing in the
   server log